chore: separate integrations for dropIn and componnts
