+++ Demo/SwiftUI/SceneDelegate.swift
@@ -19,7 +19,7 @@ internal class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         // session are new (see `application:configurationForConnectingSceneSession` instead).
 
         // Create the SwiftUI view that provides the window contents.
-        let contentView = ContentView()
+        let contentView = ComponentsView()
 
         // Use a UIHostingController as window root view controller.
         if let windowScene = scene as? UIWindowScene {
