+++ Demo/UIKit/ComponentsViewController.swift
@@ -1,0 +1,63 @@
+//
+// Copyright (c) 2020 Adyen N.V.
+//
+// This file is open source and available under the MIT license. See the LICENSE file for more info.
+//
+
+import Adyen
+
+internal final class ComponentsViewController: UIViewController  {
+    
+    private lazy var componentsView = ComponentsView()
+    private lazy var model: ComponentsModel = {
+        let model = ComponentsModel()
+        model.presenter = self
+        return model
+    }()
+    
+    // MARK: - View
+    
+    override internal func loadView() {
+        view = componentsView
+    }
+    
+    override internal func viewDidLoad() {
+        super.viewDidLoad()
+        navigationItem.title = "Components"
+        
+        componentsView.items = model.items
+        model.requestPaymentMethods()
+    }
+
+}
+
+extension ComponentsViewController: Presenter {
+
+    internal func presentAlert(with error: Error, retryHandler: (() -> Void)? = nil) {
+        let alertController = UIAlertController(title: "Error", message: error.localizedDescription, preferredStyle: .alert)
+
+        if let retryHandler = retryHandler {
+            alertController.addAction(UIAlertAction(title: "Retry", style: .default, handler: { _ in
+                retryHandler()
+            }))
+        } else {
+            alertController.addAction(UIAlertAction(title: "OK", style: .default, handler: nil))
+        }
+
+        adyen.present(alertController, animated: true)
+    }
+
+    internal func presentAlert(withTitle title: String) {
+        let alertController = UIAlertController(title: title, message: nil, preferredStyle: .alert)
+        alertController.addAction(UIAlertAction(title: "OK", style: .default, handler: nil))
+        adyen.present(alertController, animated: true)
+    }
+
+    internal func present(viewController: UIViewController, completion: (() -> Void)?) {
+        present(viewController, animated: true, completion: completion)
+    }
+
+    internal func dismiss(completion: (() -> Void)?) {
+        dismiss(animated: true, completion: completion)
+    }
+}
